Folder restructure and PyPi module made
