+++ vulcanai/utils.py
